+++ NEWS.md
@@ -1,5 +1,7 @@
 # MiscMetabar 0.33 (in development)
 
+* Add function `plot_tax_pq()` to plot taxonomic distribution (nb of sequences or nb of ASV) across factor.
+
 
 # MiscMetabar 0.32